changed code format - qt standard, merged Audio
source files
