+++ CMakeLists.txt
@@ -8,11 +8,11 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(PKG_CONFIG_USE_CMAKE_PREFIX_PATH ON)
 find_package(PkgConfig REQUIRED)
 if ( NOT (PKGCONFIG_FOUND))
-      message(FATAL_ERROR "Please Install PPkgConfig: CMake will Exit")
+    message(FATAL_ERROR "Please Install PPkgConfig: CMake will Exit")
 endif()
 pkg_check_modules(GST REQUIRED gstreamer-1.0>=1.2)
 if ( NOT (GST_FOUND))
-      message(FATAL_ERROR "Please Install Gstreamer Dev: CMake will Exit")
+    message(FATAL_ERROR "Please Install Gstreamer Dev: CMake will Exit")
 endif()
 set(ENV{PKG_CONFIG_PATH})
 
@@ -25,19 +25,18 @@ include_directories(${enginePath}/include)
 set (engineSrc ${enginePath}/source)
 set(SOURCES
     ${enginePath}/PiRadioApp.cpp
-    ${engineSrc}/AudioEngine.cpp
-    ${engineSrc}/AudioManager.cpp
+    ${engineSrc}/Audio.cpp
     ${engineSrc}/Station.cpp
     ${engineSrc}/dbStationTxt.cpp
     ${engineSrc}/CommEngine.cpp
     ${engineSrc}/RadioManager.cpp
     ${engineSrc}/Logs.cpp
-)
+    )
 
 add_executable(${PROJECT_NAME} ${SOURCES})
 target_link_libraries(
-    ${PROJECT_NAME} 
+    ${PROJECT_NAME}
     ${GST_LIBRARIES}
     zmq
     pthread
-)
+    )
