+++ engine/include/Database.hpp
@@ -4,19 +4,21 @@
 
 #include "Station.hpp"
 
-class Database {
+namespace db {
+    class Database {
+    public:
+        virtual ~Database(){}
 
-public:
-    virtual ~Database(){}
+        virtual bool isLoad() = 0;
+        virtual void load() = 0;
+        virtual void save() = 0;
 
-    virtual bool isLoad() = 0;
-    virtual void load() = 0;
-    virtual void save() = 0;
+        virtual Station* getByName(std::string name) = 0;
+        virtual Station* getByUri(std::string uri) = 0;
+        virtual void put(Station* station) = 0;
+        virtual void* getDatabase() = 0;
+        virtual void remove(Station* station) = 0;
+        virtual std::string getNamesInString() = 0;
+    };
 
-    virtual Station* getByName(std::string name) = 0;
-    virtual Station* getByUri(std::string uri) = 0;
-    virtual void put(Station* station) = 0;
-    virtual void* getDatabase() = 0;
-    virtual void remove(Station* station) = 0;
-    virtual std::string getNamesInString() = 0;
-};
+}
