+++ engine/include/RadioManager.hpp
@@ -6,22 +6,22 @@
 #include "Audio.hpp"
 #include "Logs.hpp"
 
-class RadioManager {
+class RadioManager
+{
 
-    Database* database;
-    audio::Manager* manager;
-    comm::Engine* communication;
-    
+    db::Database *database;
+    audio::Manager *manager;
+    comm::Engine *communication;
+
 public:
-    RadioManager(   Database* database,
-                    audio::Manager* audioEngineManager,
-                    comm::Engine* communication);
+    RadioManager(db::Database *database,
+                 audio::Manager *audioEngineManager,
+                 comm::Engine *communication);
     ~RadioManager();
 
-    void setDatabase(Database* database);
-    void setAudio(audio::Manager* audioEngineManager);
-    void setCommunication(comm::Engine* communication);
+    void setDatabase(db::Database *database);
+    void setAudio(audio::Manager *audioEngineManager);
+    void setCommunication(comm::Engine *communication);
     std::string execute(std::vector<std::string> command);
     void start();
-
-};  
+};
