+++ engine/include/Station.hpp
@@ -2,25 +2,25 @@
 
 #include <string>
 
-class Station {
+class Station
+{
 
     std::string name = "";
     std::string uri = "";
 
-    public:
+public:
+    Station(std::string name, std::string uri);
 
-        Station(std::string name, std::string uri);
+    std::string getName();
+    std::string getUri();
 
-        std::string getName();
-        std::string getUri();
+    void setName(std::string name);
+    void setUri(std::string uri);
 
-        void setName(std::string name);
-        void setUri(std::string uri);
+    std::string toString();
 
-        std::string toString();
-
-        static bool check (Station* station);
-        static std::string getString(Station* station);
-        static std::string getName(Station* station);
-        static std::string getUri(Station* station);
+    static bool check(Station *station);
+    static std::string getString(Station *station);
+    static std::string getName(Station *station);
+    static std::string getUri(Station *station);
 };
