+++ engine/source/Station.cpp
@@ -33,7 +33,7 @@ bool Station::check(Station* station)
 {
     if (station == nullptr)
         return false;
-    else 
+    else
         return true;
 }
 
@@ -42,7 +42,7 @@ std::string Station::getString(Station* station)
     std::string reply = "";
     if (check(station))
         reply = station->toString();
-    else 
+    else
         reply = "Error StationDoesNotExist";
     return reply;
 }
@@ -51,7 +51,7 @@ std::string Station::getName(Station* station)
 {
     if (check(station))
         return station->getName();
-    else 
+    else
         return "StationDoesNotExist";
 }
 
@@ -59,6 +59,6 @@ std::string Station::getUri(Station* station)
 {
     if (check(station))
         return station->getUri();
-    else 
+    else
         return "StationDoesNotExist";
 }
